+++ internal/modules/rol/service.go
@@ -3,8 +3,8 @@ package rol
 type RolService interface {
 	CreateRol(rol *Rol) error
 	GetAll() ([]*Rol, error)
-	GetByID(id int64) (*Rol, error)
-	Delete(id int64) (string, error)
+	GetByID(id uint) (*Rol, error)
+	Delete(id uint) (string, error)
 	Update(rol *Rol) (string, error)
 }
 
@@ -24,11 +24,11 @@ func (s *rolService) GetAll() ([]*Rol, error) {
 	return s.repo.GetAll()
 }
 
-func (s *rolService) GetByID(id int64) (*Rol, error) {
+func (s *rolService) GetByID(id uint) (*Rol, error) {
 	return s.repo.GetByID(id)
 }
 
-func (s *rolService) Delete(id int64) (string, error) {
+func (s *rolService) Delete(id uint) (string, error) {
 	return s.repo.Delete(id)
 }
